Move implementation classes to corresponding packages
