+++ core/src/main/scala/react/base/package.scala
